Комментаирии + рефакторинг
