+++ MathCore.sln.DotSettings
@@ -47,6 +47,7 @@
 	<s:Boolean x:Key="/Default/UserDictionary/Words/=Unsubscriber/@EntryIndexedValue">True</s:Boolean>
 	<s:Boolean x:Key="/Default/UserDictionary/Words/=Vetterling/@EntryIndexedValue">True</s:Boolean>
 	<s:Boolean x:Key="/Default/UserDictionary/Words/=Ziggurat/@EntryIndexedValue">True</s:Boolean>
+	<s:Boolean x:Key="/Default/UserDictionary/Words/=_0410_043F_043F_0440_043E_043A_0441_0438_043C_0430_0442_043E_0440/@EntryIndexedValue">True</s:Boolean>
 	<s:Boolean x:Key="/Default/UserDictionary/Words/=_0410_0440_043A_043A_0430_0442_0430_043D_0433_0435_043D_0441/@EntryIndexedValue">True</s:Boolean>
 	<s:Boolean x:Key="/Default/UserDictionary/Words/=_0411_0435_0440_043D_0448_0442_0435_0439_043D_0430/@EntryIndexedValue">True</s:Boolean>
 	<s:Boolean x:Key="/Default/UserDictionary/Words/=_0412_0438_043A_0438_043F_0435_0434_0438_044F/@EntryIndexedValue">True</s:Boolean>
@@ -70,6 +71,7 @@
 	<s:Boolean x:Key="/Default/UserDictionary/Words/=_0422_0440_0430_043D_0441_0432_0435_043A_0446_0438_044F/@EntryIndexedValue">True</s:Boolean>
 	<s:Boolean x:Key="/Default/UserDictionary/Words/=_0422_0440_0430_043D_0441_0432_0435_0446_0438_0440_0443_0435_043C_0430_044F/@EntryIndexedValue">True</s:Boolean>
 	<s:Boolean x:Key="/Default/UserDictionary/Words/=_0425_043E_0430_0440_0430/@EntryIndexedValue">True</s:Boolean>
+	<s:Boolean x:Key="/Default/UserDictionary/Words/=_042D_0440_0430_0442_043E_0441_0444_0435_043D/@EntryIndexedValue">True</s:Boolean>
 	<s:Boolean x:Key="/Default/UserDictionary/Words/=_0431_0435_0437_043F_0430_0440_0430_043C_0435_0442_0440_0438_0447_0435_0441_043A_043E_0433_043E/@EntryIndexedValue">True</s:Boolean>
 	<s:Boolean x:Key="/Default/UserDictionary/Words/=_0431_0438_0441_0435_043A_0446_0438_0438/@EntryIndexedValue">True</s:Boolean>
 	<s:Boolean x:Key="/Default/UserDictionary/Words/=_0432_0435_0440_0445_043D_0435/@EntryIndexedValue">True</s:Boolean>
@@ -120,6 +122,7 @@
 	<s:Boolean x:Key="/Default/UserDictionary/Words/=_0441_0435_0440_0438_0430_043B_0438_0437_0443_0435_043C_044B_0439/@EntryIndexedValue">True</s:Boolean>
 	
 	<s:Boolean x:Key="/Default/UserDictionary/Words/=_0442_0430_0439_043C_0430_0443_0442/@EntryIndexedValue">True</s:Boolean>
+	<s:Boolean x:Key="/Default/UserDictionary/Words/=_0442_0430_0439_043C_0430_0443_0442_0430/@EntryIndexedValue">True</s:Boolean>
 	<s:Boolean x:Key="/Default/UserDictionary/Words/=_0443_0432_0435_0434_043E_043C_043B_044F_0435_043C_044B_0439/@EntryIndexedValue">True</s:Boolean>
 	<s:Boolean x:Key="/Default/UserDictionary/Words/=_0444_043E_0440_043C_0430_0442_0442_0435_0440/@EntryIndexedValue">True</s:Boolean></wpf:ResourceDictionary>
