+++ Tests/ConsoleTest/Program.cs
@@ -1,4 +1,6 @@
 ﻿using System;
+using System.Runtime.Serialization;
+using Microsoft.Data.Analysis;
 
 namespace ConsoleTest
 {
@@ -6,9 +8,17 @@ namespace ConsoleTest
     {
         private static void Main(string[] args)
         {
-            var q = '0'.ToDigit();
+            var times = new PrimitiveDataFrameColumn<DateTime>("Date");
+            times.Append(DateTime.Now.Subtract(TimeSpan.FromDays(2)));
+            times.Append(DateTime.Now.Subtract(TimeSpan.FromDays(1)));
+            times.Append(DateTime.Now);
 
-            Console.WriteLine(q);
+            var ints = new PrimitiveDataFrameColumn<int>("ints", 3);
+            var strs = new StringDataFrameColumn("strings", 3);
+
+            var data = new DataFrame(times, ints, strs);
+
+            data[0, 1] = 10;
 
             Console.ReadLine();
         }
